feat: move About+Support to Advanced tab, fancy Ko-fi button
- About (version, source, license) and Support development sections
  moved from Self-hosted tab to bottom of Advanced tab. Less cluttered
  Self-hosted tab; About is universal so it belongs with shared config.
- Ko-fi button reskinned: gradient red background, Lucide coffee icon,
  hover lift effect. Inline anchor + setIcon (no innerHTML, no external
  image load).
