+++ src/tabs/self-hosted-tab.ts
@@ -146,36 +146,4 @@ export function renderSelfHostedTab(ctx: TabContext): void {
 				});
 			});
 	}
-
-	// ── About ──
-	new Setting(containerEl).setName("About").setHeading();
-
-	const aboutList = containerEl.createEl("ul", { cls: "engram-about-list" });
-
-	const versionItem = aboutList.createEl("li");
-	versionItem.createSpan({ text: "Version: " });
-	versionItem.createSpan({ text: plugin.manifest.version });
-
-	const repoItem = aboutList.createEl("li");
-	repoItem.createSpan({ text: "Source: " });
-	repoItem.createEl("a", {
-		text: "github.com/Rasbandit/Engram-obsidian-sync",
-		href: "https://github.com/Rasbandit/Engram-obsidian-sync",
-	});
-
-	const licenseItem = aboutList.createEl("li");
-	licenseItem.createSpan({ text: "License: MIT" });
-
-	new Setting(containerEl)
-		.setName("Support development")
-		.setDesc(
-			"If this plugin saves you time, consider supporting development. Optional and appreciated.",
-		)
-		.addButton((btn) => {
-			btn.setButtonText("Ko-fi")
-				.setCta()
-				.onClick(() => {
-					window.open("https://ko-fi.com/rasbandit", "_blank");
-				});
-		});
 }
